move more lessons to search and hardcode featured lesson
