+++ site/src/app/(browse)/home/page.tsx
@@ -10,8 +10,8 @@ import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
 import { ArrowRight, Brain, Star, Trophy } from 'lucide-react';
 import { BrowseHeader } from '@/components/typography/BrowseHeader';
-import { GengarIcon } from '@/components/icons/gengar-icon';
 import Image from 'next/image';
+import Link from 'next/link';
 
 export default async function HomePage() {
   return (
@@ -44,9 +44,6 @@ export default async function HomePage() {
 
       {/* Featured Lesson */}
       <FeaturedLesson />
-
-      {/* More Lessons Grid */}
-      <MoreLessons />
     </div>
   );
 }
@@ -90,110 +87,25 @@ const FeaturedLesson = () => (
           Featured
         </Badge>
       </div>
-      <CardTitle className="text-xl sm:text-2xl">
-        Dynamic Programming Basics
-      </CardTitle>
+      <CardTitle className="text-xl sm:text-2xl">Grind 75</CardTitle>
       <CardDescription className="text-base">
-        Master the fundamentals of DP with 5 carefully selected problems
+        Go through the most important questions, inspired by the "Grind 75"
+        list.
       </CardDescription>
     </CardHeader>
     <CardContent>
       <div className="flex flex-col sm:flex-row gap-4 items-start sm:items-center justify-between">
         <div className="flex items-center gap-2 text-sm text-muted-foreground">
           <Brain className="h-4 w-4" />
-          <span>5 Problems • 2 hours</span>
-          <div className="flex items-center gap-1">
-            <Star className="h-4 w-4 fill-yellow-500 text-yellow-500" />
-            <span>4.9</span>
-          </div>
+          <span>75 Problems • 2 hours</span>
         </div>
-        <Button className="w-full sm:w-auto">
-          Start Lesson
-          <ArrowRight className="ml-2 h-4 w-4" />
-        </Button>
+        <Link href="/study-session">
+          <Button className="w-full sm:w-auto">
+            Start Lesson
+            <ArrowRight className="ml-2 h-4 w-4" />
+          </Button>
+        </Link>
       </div>
     </CardContent>
   </Card>
-);
-
-// TODO: Make this a client component with more enriched data
-const MoreLessons = () => (
-  <div className="space-y-4">
-    <div className="flex items-center justify-between">
-      <h2 className="text-xl font-bold tracking-tight">More Lessons</h2>
-      <Button variant="ghost" className="text-sm">
-        View all
-      </Button>
-    </div>
-    <div className="grid gap-4 sm:grid-cols-2 lg:grid-cols-3">
-      <Card className="flex flex-col">
-        <CardHeader>
-          <Badge variant="destructive" className="w-fit">
-            Hard
-          </Badge>
-          <CardTitle className="mt-2">Graph Algorithms</CardTitle>
-          <CardDescription>
-            Advanced graph problems focusing on shortest paths
-          </CardDescription>
-        </CardHeader>
-        <CardContent className="flex-1">
-          <div className="flex items-center gap-2 text-sm text-muted-foreground">
-            <Brain className="h-4 w-4" />
-            <span>4 Problems • 3h</span>
-          </div>
-        </CardContent>
-        <div className="p-4 pt-0">
-          <Button variant="outline" className="w-full">
-            Start Lesson
-          </Button>
-        </div>
-      </Card>
-
-      <Card className="flex flex-col">
-        <CardHeader>
-          <Badge variant="secondary" className="w-fit">
-            New
-          </Badge>
-          <CardTitle className="mt-2">Binary Trees</CardTitle>
-          <CardDescription>
-            Common patterns for solving tree-based problems
-          </CardDescription>
-        </CardHeader>
-        <CardContent className="flex-1">
-          <div className="flex items-center gap-2 text-sm text-muted-foreground">
-            <Brain className="h-4 w-4" />
-            <span>6 Problems • 2.5h</span>
-          </div>
-        </CardContent>
-        <div className="p-4 pt-0">
-          <Button variant="outline" className="w-full">
-            Start Lesson
-          </Button>
-        </div>
-      </Card>
-
-      <Card className="flex flex-col sm:hidden lg:flex">
-        <CardHeader>
-          <Badge variant="secondary" className="w-fit">
-            Popular
-          </Badge>
-          <CardTitle className="mt-2">Array Techniques</CardTitle>
-          <CardDescription>
-            Essential array manipulation strategies
-          </CardDescription>
-        </CardHeader>
-        <CardContent className="flex-1">
-          <div className="flex items-center gap-2 text-sm text-muted-foreground">
-            <Brain className="h-4 w-4" />
-            <span>5 Problems • 2h</span>
-          </div>
-        </CardContent>
-        <div className="p-4 pt-0">
-          <Button variant="outline" className="w-full">
-            Start Lesson
-          </Button>
-        </div>
-      </Card>
-    </div>
-  </div>
 );
